+++ spec/ey_yml_customized_deploy_spec.rb
@@ -45,6 +45,10 @@ describe "Deploying an app with ey.yml" do
     it "makes custom variables available to hooks" do
       @deploy_dir.join('current', 'custom_hook').read.should include("custom_from_ey_yml")
     end
+
+    it "doesn't display the database adapter warning with ignore_database_adapter_warning: true" do
+      read_output.should_not =~ /WARNING/
+    end
   end
 
   context "with a different ey.yml" do
@@ -55,6 +59,10 @@ describe "Deploying an app with ey.yml" do
     it "always installs maintenance pages" do
       @deploy_dir.join('current','maintenance_enabled').should exist
       @deploy_dir.join('current','maintenance_disabled').should_not exist
+    end
+
+    it "displays the database adapter warning without ignore_database_adapter_warning" do
+      read_output.should =~ /WARNING: Gemfile.lock does not contain a recognized database adapter./
     end
   end
 end